+++ src/EndPoints/Endpoint.php
@@ -2,16 +2,11 @@
 
 namespace seregazhuk\HeadHunterApi\EndPoints;
 
-use seregazhuk\HeadHunterApi\Contracts\RequestInterface;
+use seregazhuk\HeadHunterApi\Request;
 
 abstract class Endpoint
 {
     const RESOURCE = '/resource';
-
-    /**
-     * @var RequestInterface
-     */
-    protected $request;
 
     /**
      * @var EndpointsContainer
@@ -24,7 +19,6 @@ abstract class Endpoint
     public function __construct(EndpointsContainer $container)
     {
         $this->container = $container;
-        $this->request = $container->getRequest();
     }
 
     /**
@@ -86,22 +80,28 @@ abstract class Endpoint
         $this->requestResource('delete', $verb);
     }
 
+    /**
+     * @param string $method
+     * @param string $verb
+     * @param array $params
+     * @return mixed
+     */
     protected function requestResource($method = 'get', $verb = '', $params = [])
     {
         $method = strtolower($method);
 
-        return $this->request->makeRequest(
-            $method, $this->getResourceUri($verb), $params
-        );
+        return $this->callRequestMethod($method, $verb, $params);
     }
 
+    /**
+     * @param string $method
+     * @param string $verb
+     * @param array $params
+     * @return mixed
+     */
     protected function requestResourceJson($method = 'get', $verb = '', $params = [])
     {
-        $method = strtolower($method);
-
-        return $this->request->makeRequest(
-            $method, $this->getResourceUri($verb), $params
-        );
+        return $this->requestResource($method. 'Json', $verb, $params);
     }
 
     /**
@@ -116,10 +116,25 @@ abstract class Endpoint
     }
 
     /**
-     * @return RequestInterface
+     * @return Request
      */
     protected function getRequest()
     {
         return $this->request;
+    }
+
+    /**
+     * @param $method
+     * @param $verb
+     * @param $params
+     * @return mixed
+     */
+    protected function callRequestMethod($method, $verb, $params)
+    {
+        $request = $this->container->getRequest();
+
+        return $request->$method(
+            $this->getResourceUri($verb), $params
+        );
     }
 }