+++ src/Request.php
@@ -2,67 +2,132 @@
 
 namespace seregazhuk\HeadHunterApi;
 
+use GuzzleHttp\Client;
+use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\Psr7\Request as GuzzleRequest;
 
-use seregazhuk\HeadHunterApi\Contracts\HttpInterface;
-use seregazhuk\HeadHunterApi\Contracts\RequestInterface;
-use seregazhuk\HeadHunterApi\Exceptions\HeadHunterApiException;
-
-class Request implements RequestInterface
+class Request
 {
     /**
-     * @var HttpInterface
+     * @var \GuzzleHttp\Client
      */
     protected $client;
 
-    public function __construct(HttpInterface $http, $token = null)
-    {
-        $this->client = $http;
+    /**
+     * @var array
+     */
+    protected $headers = [];
 
-        if($token) $this->client->setHeaders(['Authorization' => 'Bearer ' . $token]);
+    public function __construct($baseUrl, $token = null)
+    {
+        $this->client = new Client(['base_uri' => $baseUrl]);
+
+        if($token) $this->setHeaders(['Authorization' => 'Bearer ' . $token]);
     }
 
     /**
-     * @param string $requestMethod
      * @param string $uri
      * @param array $params
-     * @param bool $useJson
-     * @return mixed
-     * @throws HeadHunterApiException
+     * @return array|null
      */
-    public function makeRequest($requestMethod, $uri, $params = [], $useJson = false)
+    public function get($uri, $params = [])
     {
-        $requestMethod = strtolower($requestMethod);
-
-        return $this->executeRequest($requestMethod, $uri, $params);
-    }
-
-    /**
-     * @param string $requestMethod
-     * @param string $uri
-     * @param array $params
-     * @return mixed
-     * @throws HeadHunterApiException
-     */
-    public function makeJsonRequest($requestMethod, $uri, $params = [])
-    {
-        $requestMethod = strtolower($requestMethod) . 'json';
-
-        return $this->executeRequest($requestMethod, $uri, $params);
-    }
-
-    /**
-     * @param $requestMethod
-     * @param $uri
-     * @param array $params
-     * @return mixed
-     * @throws HeadHunterApiException
-     */
-    protected function executeRequest($requestMethod, $uri, array $params = [])
-    {
-        if(!method_exists($this->client, $requestMethod)) {
-            throw new HeadHunterApiException("Request method $requestMethod not found");
+        if(!empty($params)){
+            $uri .= '?'. http_build_query($params);
         }
 
-        return $this->client->$requestMethod($uri, $params);
+        return $this->executeRequest('GET', $uri);
+    }
+
+    /**
+     * @param string $uri
+     * @param array $params
+     * @return array|null
+     */
+    public function post($uri, $params = [])
+    {
+        return $this->executeRequest(
+            'POST', $uri, ['query' => $params]
+        );
+    }
+
+    /**
+     * @param string $uri
+     * @param array $params
+     * @return array|null
+     */
+    public function postJson($uri, $params = [])
+    {
+        return $this->executeRequest(
+            'POST', $uri, ['json' => $params]
+        );
+    }
+
+    /**
+     * @param string $uri
+     * @param array $params
+     * @return array|null
+     */
+    public function put($uri, $params = [])
+    {
+        return $this->executeRequest(
+            'PUT', $uri, ['query' => $params]
+        );
+    }
+
+    /**
+     * @param string $uri
+     * @param array $params
+     * @return array|null
+     */
+    public function putJson($uri, $params = [])
+    {
+        return $this->executeRequest(
+            'PUT', $uri, ['json' => $params]
+        );
+    }
+
+    /**
+     * @param string $uri
+     * @return array|null
+     */
+    public function delete($uri)
+    {
+        return $this->executeRequest('DELETE', $uri);
+    }
+
+    /**
+     * @param Response $response
+     * @return array|null
+     */
+    private function parseResponse(Response $response)
+    {
+        return json_decode($response->getBody(), true);
+    }
+
+    /**
+     * @param string $method
+     * @param string $uri
+     * @param array $options
+     * @return array|null
+     */
+    protected function executeRequest($method, $uri, array $options = [])
+    {
+        $request = new GuzzleRequest($method, $uri, $this->headers);
+
+        $response = $this->client->send($request, $options);
+
+        return $this->parseResponse($response);
+    }
+
+    /**
+     * @param mixed $headers
+     * @return $this
+     */
+    public function setHeaders($headers)
+    {
+        $this->headers = $headers;
+
+        return $this;
     }
 }