+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## v0.2.2 - ??
 ### Added
- - Resume edit
+ - Resume edit, create
 
 ## v0.2.1 - 2017-04-24
 ### Changed: